# Optimize bit reversal