--- a/util/Cargo.toml
+++ b/util/Cargo.toml
@@ -6,3 +6,6 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 serde.workspace = true
+
+[dev-dependencies]
+rand.workspace = true