--- a/util/Cargo.toml
+++ b/util/Cargo.toml
@@ -9,3 +9,8 @@ serde.workspace = true
 
 [dev-dependencies]
 rand.workspace = true
+criterion.workspace = true
+
+[[bench]]
+name = "bit_reverse"
+harness = false