--- a/util/src/lib.rs
+++ b/util/src/lib.rs
@@ -163,7 +163,9 @@ fn reverse_slice_index_bits_small<F>(vals: &mut [F], lb_n: usize) {
         for src in 0..vals.len() {
             let dst = (BIT_REVERSE_6BIT[src] as usize).wrapping_shr(dst_shr_amt);
             if src < dst {
-                vals.swap(src, dst);
+                unsafe {
+                    core::ptr::swap(vals.get_unchecked_mut(src), vals.get_unchecked_mut(dst));
+                }
             }
         }
     } else {
@@ -181,7 +183,9 @@ fn reverse_slice_index_bits_small<F>(vals: &mut [F], lb_n: usize) {
                 let src = src_hi + src_lo;
                 let dst = dst_hi + dst_lo;
                 if src < dst {
-                    vals.swap(src, dst);
+                    unsafe {
+                        core::ptr::swap(vals.get_unchecked_mut(src), vals.get_unchecked_mut(dst));
+                    }
                 }
             }
         }
@@ -194,7 +198,9 @@ fn reverse_slice_index_bits_small<F>(vals: &mut [F], lb_n: usize) {
     for src in 0..vals.len() {
         let dst = src.reverse_bits().wrapping_shr(usize::BITS - lb_n as u32);
         if src < dst {
-            vals.swap(src, dst);
+            unsafe {
+                core::ptr::swap(vals.get_unchecked_mut(src), vals.get_unchecked_mut(dst));
+            }
         }
     }
 }