@unzvfu 2025-01-08
> As the benches are mitigated on performance (considering the priority of x86_64), I'm happy to revert to the more Pareto-efficient https://github.com/Plonky3/Plonky3/commit/1419ae529dc77decad50cfa70e60afbbf72af9f4.

Regarding this, if we must choose, I think we should err towards whatever is fastest on `x86_64` for now. It would be nice to have the fastest times on both architectures, but that would probably complicate the implementation more than is really worthwhile, since none of these functions are bottlenecks iirc.