@huitseeker 2025-01-09
Rebased, re-added comments, and re-added calls to unchecked. The performance picture is quite similar.

<details><Summary> Benchmarks on 373bbd3 (toggle me!)</summary>

**on x86_64 (Ryzen 9 9950X)**:
```
reverse_bits/2          time:   [556.12 ps 556.28 ps 556.53 ps]
                        change: [+0.3869% +0.4483% +0.5264%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_bits/8          time:   [556.39 ps 556.42 ps 556.45 ps]
                        change: [+0.3291% +0.3448% +0.3573%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_bits/32         time:   [556.47 ps 556.52 ps 556.58 ps]
                        change: [+4.8394% +4.8591% +4.8778%] (p = 0.00 < 0.05)
                        Performance has regressed.
reverse_bits/256        time:   [557.27 ps 557.61 ps 558.03 ps]
                        change: [+0.4748% +0.5277% +0.5756%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_bits/65536      time:   [556.22 ps 556.26 ps 556.30 ps]
                        change: [+0.2791% +0.3012% +0.3195%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_bits/16777216   time:   [556.32 ps 556.36 ps 556.41 ps]
                        change: [+5.2723% +5.3090% +5.3383%] (p = 0.00 < 0.05)
                        Performance has regressed.

reverse_slice_index_bits/2
                        time:   [7.7353 ns 7.7552 ns 7.7753 ns]
                        change: [-1.3058% -0.9487% -0.6066%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_slice_index_bits/8
                        time:   [9.0896 ns 9.1187 ns 9.1483 ns]
                        change: [-7.5568% -7.2836% -7.0244%] (p = 0.00 < 0.05)
                        Performance has improved.
reverse_slice_index_bits/32
                        time:   [14.322 ns 14.335 ns 14.348 ns]
                        change: [-27.474% -27.316% -27.173%] (p = 0.00 < 0.05)
                        Performance has improved.
reverse_slice_index_bits/256
                        time:   [99.739 ns 99.780 ns 99.832 ns]
                        change: [-42.822% -42.588% -42.352%] (p = 0.00 < 0.05)
                        Performance has improved.
reverse_slice_index_bits/65536
                        time:   [191.90 µs 191.97 µs 192.06 µs]
                        change: [-6.4595% -6.3642% -6.2548%] (p = 0.00 < 0.05)
                        Performance has improved.
reverse_slice_index_bits/16777216
                        time:   [271.77 ms 272.23 ms 272.69 ms]
                        change: [+65.384% +65.769% +66.136%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

**on aarch64 (M3 Max)**:
```
reverse_bits/2          time:   [635.06 ps 642.43 ps 649.39 ps]
                        change: [-5.3831% -4.3877% -3.2902%] (p = 0.00 < 0.05)
                        Performance has improved.
reverse_bits/8          time:   [643.66 ps 651.13 ps 657.24 ps]
                        change: [-2.2598% -1.5755% -0.8841%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_bits/32         time:   [644.55 ps 652.55 ps 660.07 ps]
                        change: [-2.4090% -1.5073% -0.5039%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_bits/256        time:   [610.10 ps 622.17 ps 633.72 ps]
                        change: [-0.5575% +1.8662% +4.0999%] (p = 0.12 > 0.05)
                        No change in performance detected.
reverse_bits/65536      time:   [610.38 ps 622.11 ps 633.73 ps]
                        change: [+0.0314% +2.2721% +4.6853%] (p = 0.06 > 0.05)
                        No change in performance detected.
reverse_bits/16777216   time:   [602.76 ps 616.13 ps 627.97 ps]
                        change: [-4.6680% -2.5346% -0.3219%] (p = 0.03 < 0.05)
                        Change within noise threshold.

reverse_slice_index_bits/2
                        time:   [15.520 ns 15.565 ns 15.614 ns]
                        change: [+0.7651% +1.2620% +1.7197%] (p = 0.00 < 0.05)
                        Change within noise threshold.
reverse_slice_index_bits/8
                        time:   [21.077 ns 21.135 ns 21.196 ns]
                        change: [-1.1586% -0.6395% -0.1422%] (p = 0.02 < 0.05)
                        Change within noise threshold.
reverse_slice_index_bits/32
                        time:   [33.484 ns 33.694 ns 33.926 ns]
                        change: [+2.1416% +2.9553% +3.7610%] (p = 0.00 < 0.05)
                        Performance has regressed.
reverse_slice_index_bits/256
                        time:   [149.33 ns 150.43 ns 151.56 ns]
                        change: [-0.7738% -0.1035% +0.5201%] (p = 0.76 > 0.05)
                        No change in performance detected.
reverse_slice_index_bits/65536
                        time:   [122.04 µs 122.40 µs 122.81 µs]
                        change: [-17.413% -17.029% -16.649%] (p = 0.00 < 0.05)
                        Performance has improved.
reverse_slice_index_bits/16777216
                        time:   [62.282 ms 62.579 ms 62.891 ms]
                        change: [-46.730% -46.177% -45.613%] (p = 0.00 < 0.05)
                        Performance has improved.
```

</details>