@unzvfu 2025-01-13
> Rebased, re-added comments, and re-added calls to unchecked. The performance picture is quite similar.
> Benchmarks on [373bbd3](https://github.com/Plonky3/Plonky3/commit/373bbd33c7a8c75467be0a51b06b0511b4313d6a) (toggle me!)

Okay, thanks for that. Since using unchecked doesn't affect performance I agree that your use of `swap` is preferable. I think this will be ready to merge once you revert back to using `swap`, and (as mentioned previously) revert to the version that is fastest on x86.